Add tests for the ratings report
